requirements.txt
streamlit==1.40.2
numpy==2.4.0
pillow==10.4.0
lark==1.2.2
streamlit-autorefresh==1.0.1
plotly==5.24.1
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
